requirements.txt
aiogram
fastapi
uvicorn
sqlalchemy
asyncpg         # для PostgreSQL (можно заменить на sqlite, если нужно)
python-dotenv
pybit
openai
cryptography    # для шифрования ключей
